+++ CHANGELOG.md
@@ -3,10 +3,11 @@ All notable changes to this project will be documented in this file. This change
 
 ## 0.1.1 - 2020-01-29
 ### Added
-* `update`, `field-delete`, `increment`, `server-timestamp`, `array-union`, `array-remove`, `get-id`
+* `merge!`, `assoc!`, `dissoc!`, `field-delete`, `inc`, `server-timestamp`, `array-union`, `array-remove`, `id`
 
 ### Modified
 - `->atom` now returns an atom instead of a promise.
+- Renamed `add` -> `add!`, `set` -> `set!`, `delete` -> `delete!`
 
 ## 0.1.0 - 2020-01-28
 ### Added